requirements.txt
altgraph==0.17.4
attrs==25.3.0
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
coremltools==8.2
cv2-enumerate-cameras==1.1.18.3
cycler==0.12.1
cython==0.29.37
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
glfw==2.8.0
idna==3.10
imgui==2.0.0
iniconfig==2.1.0
jinja2==3.1.6
kiwisolver==1.4.8
lite-camera==2.0.4
macholib==1.16.3
markupsafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
pluggy==1.5.0
protobuf==6.30.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyaml==25.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyopengl==3.1.9
pyparsing==3.2.1
pysdl2==0.9.17
pysdl2-dll==2.30.10
pysimplegui==5.0.8.2
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32-ctypes==0.2.3
pyyaml==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==76.1.0
six==1.17.0
sympy==1.13.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.67.1
tracker==0.1.1
typing-extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.93
ultralytics-thop==2.0.14
urllib3==2.3.0
